pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-detector"
version = "0.1.0"
description = "Fake news detection toolkit"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "nltk>=3.8",
    "joblib>=1.3",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "shap>=0.44",
    "transformers>=4.36",
    "datasets>=2.15",
    "torch>=2.1",
    "streamlit>=1.29",
    "typer>=0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=23.12",
    "ruff>=0.1.7"
]

[tool.setuptools.packages.find]
where = ["src"]
